+++ config.py
@@ -2,11 +2,12 @@ import tempfile
 import json
 
 # Build Config
-delay = 35
+delay = 50
 name = "macOSBigSur"
 sizes = [24, 28, 32, 40, 48, 56, 64, 72, 80, 88, 96]
 
 bitmaps_dir = "./bitmaps"
+package_dir = "./themes"
 temp_folder = tempfile.mkdtemp()
 
 # Cleanup Configs
@@ -20,26 +21,25 @@ with open("./package.json") as f:
 
 # Windows Cursors Config
 windows_cursors = {
-    "all-scroll.cur": "move.cur",
-    "bd_double_arrow.cur": "diagonal-resize-1.cur",
-    "bottom_left_corner.cur": "diagonal-resize-2.cur",
-    "bottom_side.cur": "vertical-resize.cur",
-    "circle.cur": "unavailable.cur",
-    "crosshair.cur": "precision-select.cur",
-    "dnd-ask.cur": "help-select.cur",
-    "hand2.cur": "link-select.cur",
-    "left_ptr.cur": "normal-select.cur",
-    "left_ptr_watch.ani": "working-in-background.ani",
-    "pencil.cur": "handwriting.cur",
-    "right_side.cur": "horizontal-resize.cur",
-    "sb_up_arrow.cur": "alt-select.cur",
-    "wait.ani": "busy.ani",
-    "x_cursor.cur": "pirate.cur",
-    "xterm.cur": "text-select.cur"
+    "left_ptr_watch.ani": "AppStarting.ani",
+    "left_ptr.cur": "Arrow.cur",
+    "crosshair.cur": "Cross.cur",
+    "hand2.cur": "Hand.cur",
+    "pencil.cur": "Handwriting.cur",
+    "dnd-ask.cur": "Help.cur",
+    "xterm.cur": "IBeam.cur",
+    "circle.cur": "NO.cur",
+    "all-scroll.cur": "SizeAll.cur",
+    "bd_double_arrow.cur": "SizeNESW.cur",
+    "sb_v_double_arrow.cur": "SizeNS.cur",
+    "fd_double_arrow.cur": "SizeNWSE.cur",
+    "sb_h_double_arrow.cur": "SizeWE.cur",
+    "sb_up_arrow.cur": "UpArrow.cur",
+    "wait.ani": "Wait.ani",
 }
 
 # Windows install.inf file content
 with open("./scripts/windows.inf") as f:
     data = f.read()
-    window_install_inf = data.replace(
+    window_install_inf_content = data.replace(
         "<inject_theme_name>", name+" Cursors").replace("<inject_author_name>", author)