+++ helper.py
@@ -2,11 +2,10 @@ import shutil
 import json
 import sys
 
-from config import name, temp_folder, bitmaps_dir, win_out, x11_out, window_install_inf, windows_cursors
+from config import name, temp_folder, bitmaps_dir, win_out, x11_out, window_install_inf_content, windows_cursors, package_dir
 from os import path, listdir, rename, remove
 
 
-package_dir = "./packages"
 x11_out_dir = path.join(package_dir, x11_out)
 win_out_dir = path.join(package_dir, win_out)
 
@@ -26,7 +25,7 @@ def window_bundle() -> None:
     # creating install.inf file
     install_inf_path = path.join(win_out_dir, "install.inf")
     with open(install_inf_path, "w") as file:
-        file.write(install_inf_path)
+        file.write(window_install_inf_content)
 
 
 def init_build() -> None:
@@ -37,7 +36,7 @@ def init_build() -> None:
     with open("./package.json", "r") as package_file:
         data = json.loads(package_file.read())
         version = data['version']
-        print("⚡ Build Version %s" % version)
+        print("⚡ Apple Cursor Version %s" % version)
 
     # cleanup old packages
     if path.exists(package_dir):
@@ -60,16 +59,3 @@ def pack_it() -> None:
 
     # create install.inf file in Windows Theme
     window_bundle()
-
-    # Packaging
-    #  - .tar archive for X11
-    #  - .zip archive for Windows
-    shutil.make_archive(x11_out_dir, "tar", x11_out_dir)
-    shutil.make_archive(win_out_dir, "zip", win_out_dir)
-
-    # Clenaup
-    shutil.rmtree(temp_folder)
-    for f in listdir(package_dir):
-        f_path = path.join(package_dir, f)
-        if path.isdir(f_path):
-            shutil.rmtree(f_path)